+++ gradle.properties
@@ -2,18 +2,18 @@
 org.gradle.jvmargs=-Xmx1G
 
 # Fabric Properties
-minecraft_version=20w10a
-yarn_mappings=20w10a+build.7
-loader_version=0.7.8+build.186
-loom_version=0.2.6-SNAPSHOT
+minecraft_version=1.16.4
+yarn_mappings=1.16.4+build.7
+loader_version=0.10.8
+loom_version=0.5-SNAPSHOT
 
 # Mod Properties
-mod_version = 1.1.0+build.5
+mod_version = 1.1.0+build.6
 archives_base_name = netherHigher
 
 # Dependencies
-fabric_version=0.1.9+f3d8141be4
-clothconfig_version=3.0.1-unstable.202003051928
-modmenu_version=1.10.2+build.1
+fabric_version=0.2.8+35e08e333b
+clothconfig_version=4.8.3
+modmenu_version=1.14.13+build.22
 clientcommands_version=0.4.2+1.14.3-SNAPSHOT
 fiber_version=0.8.0-1
