+++ src/main/java/site/geni/netherhigher/gui/Categories.java
@@ -3,7 +3,8 @@ package site.geni.netherhigher.gui;
 import me.shedaniel.clothconfig2.api.ConfigCategory;
 import me.shedaniel.clothconfig2.impl.builders.BooleanToggleBuilder;
 import me.shedaniel.clothconfig2.impl.builders.TextDescriptionBuilder;
-import net.minecraft.client.resource.language.I18n;
+import net.minecraft.text.LiteralText;
+import net.minecraft.text.TranslatableText;
 import site.geni.netherhigher.NetherHigher;
 
 @SuppressWarnings("ConstantConditions")
@@ -19,9 +20,9 @@ public class Categories {
 			// Adds the warning in the "World" category
 			general.addEntry(
 				new TextDescriptionBuilder(
-					"",
-					"",
-					I18n.translate("config.netherhigher.warning")
+						new LiteralText(""),
+						new LiteralText(""),
+						new TranslatableText("config.netherhigher.warning")
 				).setColor(0xff5555)
 					.build()
 			);
@@ -29,8 +30,8 @@ public class Categories {
 			// Adds the option for raising the Nether's height limit to 256
 			general.addEntry(
 				new BooleanToggleBuilder(
-					"text.cloth.reset_value",
-					"config.netherhigher.raiseNetherHeightLimit",
+					new TranslatableText("text.cloth.reset_value"),
+					new TranslatableText("config.netherhigher.raiseNetherHeightLimit"),
 					NetherHigher.getConfig().raiseNetherHeightLimit.getValue()
 				).setDefaultValue(NetherHigher.getConfig().raiseNetherHeightLimit.getDefaultValue())
 					.setSaveConsumer(bool -> NetherHigher.getConfig().raiseNetherHeightLimit.setValue(bool))
@@ -40,8 +41,8 @@ public class Categories {
 			// Adds the option for raising the End's height limit to 256
 			general.addEntry(
 				new BooleanToggleBuilder(
-					"text.cloth.reset_value",
-					"config.netherhigher.raiseEndHeightLimit",
+					new TranslatableText("text.cloth.reset_value"),
+					new TranslatableText("config.netherhigher.raiseEndHeightLimit"),
 					NetherHigher.getConfig().raiseEndHeightLimit.getValue()
 				).setDefaultValue(NetherHigher.getConfig().raiseEndHeightLimit.getDefaultValue())
 					.setSaveConsumer(bool -> NetherHigher.getConfig().raiseEndHeightLimit.setValue(bool))
