--- src/main/java/site/geni/netherhigher/mixins/common/world/biome/DefaultBiomeFeaturesMixin.java
@@ -1,28 +1,0 @@
-package site.geni.netherhigher.mixins.common.world.biome;
-
-import net.minecraft.world.biome.CrimsonForestBiome;
-import net.minecraft.world.biome.DefaultBiomeFeatures;
-import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.injection.Constant;
-import org.spongepowered.asm.mixin.injection.ModifyConstant;
-import site.geni.netherhigher.NetherHigher;
-
-@Mixin(DefaultBiomeFeatures.class)
-public abstract class DefaultBiomeFeaturesMixin {
-	/**
-	 * Sets some nether biomes' features' height limit to 256, depending on the mod's configuration
-	 *
-	 * @param original Original integer value of 128
-	 * @return Either 256 or the default of 128, depending on the mod's configuration
-	 * @author geni
-	 */
-	@ModifyConstant(
-		constant = @Constant(
-			intValue = 128
-		),
-		method = "addNetherOres"
-	)
-	private static int setHeightLimit(int original) {
-		return NetherHigher.getConfig().raiseNetherHeightLimit.getValue() ? 256 : original;
-	}
-}
